Make improvements
- Introduce ualarm() function
- Make rename() report EISEMPTY on Windows
- Always raise EINVAL upon open(O_RDONLY|O_TRUNC)
- Add macro so ./configure will detect SOCK_CLOEXEC
- Fix O_TRUNC without O_CREAT not working on Windows
- Let fcntl(F_SETFL) change O_APPEND status on Windows
- Make sure pwrite() / pread() report ESPIPE on sockets
- Raise ESPIPE on Windows when pwrite() is used on pipe
- Properly compute O_APPEND CreateFile() flags on Windows
- Don't require O_DIRECTORY to open directories on Windows
- Fix more instances of Windows reporting EISDIR and ENOTDIR
- Normalize EFTYPE and EMLINK to ELOOP on NetBSD and FreeBSD
- Make unlink() / rmdir() work on read-only files on Windows
- Validate UTF-8 on Windows paths to fix bug with overlong NUL
- Always print signal name to stderr when crashing due to SIG_DFL
- Fix Windows bug where denormalized paths >260 chars didn't work
- Block signals on BSDs when thread exits before trashing its own stack

+++ libc/calls/weirdtypes.h
@@ -28,7 +28,7 @@ typedef int64_t register_t;
 typedef uint16_t sa_family_t; /* bsd:uint8_t */
 typedef uint32_t socklen_t;
 typedef uint32_t speed_t;
-typedef int64_t suseconds_t;   /* int32_t on xnu */
+typedef uint32_t suseconds_t;
 typedef uint64_t useconds_t;   /* uint32_t on xnu */
 typedef int64_t syscall_arg_t; /* uint64_t on xnu */
 typedef uint32_t tcflag_t;
